Pass the right key to _sign
Due to a bug in the XS code, clamping the key in keygen would also
modify the original $k in sign on perl 5.20+. This is causing failures
on older perls, and also when the variables are properly decoupled.

This uses the correct value in sign, so that it will work correctly
on all versions of perl.